Move Factory to core
