+++ tests/battle/test_battle_toon_get_attacks.py
@@ -2,16 +2,16 @@ import pytest
 
 from Battle import Battle, BattleCog, BattleToon, RewardCalculator
 from Cog import Cog
-from Factory import BattleCogFactory, CogFactory
+from core.Factory import BattleCogFactory, CogFactory
 from GagGlobals import GAG, TRACK
+from tests.fixtures.battle_fixtures import get_expected_reward
 from Toon import Toon
 from ToonGlobals import (ASTRO_EXPECTED_AVAILABLE_GAGS,
-                            ASTRO_EXPECTED_TRACK_LEVELS,
-                            ASTRO_EXPECTED_UNLOCKED_GAGS,
-                            TRAPA_EXPECTED_AVAILABLE_GAGS,
-                            TRAPA_EXPECTED_TRACK_LEVELS,
-                            TRAPA_EXPECTED_UNLOCKED_GAGS)
-from tests.fixtures.battle_fixtures import get_expected_reward
+                         ASTRO_EXPECTED_TRACK_LEVELS,
+                         ASTRO_EXPECTED_UNLOCKED_GAGS,
+                         TRAPA_EXPECTED_AVAILABLE_GAGS,
+                         TRAPA_EXPECTED_TRACK_LEVELS,
+                         TRAPA_EXPECTED_UNLOCKED_GAGS)
 
 BATTLE_ID = 1
